Add fields to subsidy report
